add Transfer, TransferDao, logic for transfer

+++ src/main/java/pk/ssi/model/Rachunek.java
@@ -45,7 +45,7 @@ public class Rachunek {
 		return saldo;
 	}
 
-	public void setSaldo(int saldo) {
+	public void setSaldo(double saldo) {
 		this.saldo = saldo;
 	}
